requirements.txt
asgiref==3.7.2
attrs==23.1.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
flatbuffers==23.5.26
idna==3.6
Markdown==3.5.1
numpy==1.26.2
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.0
pandas==2.1.4
platformdirs==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.1.0
